requirements.txt
django-allauth
psycopg2